Add user interactions
Added user interactions for the bot. Provided with "authentication" methods for both addition and removal of a user from a notification list, which is degree and year-specific.

Updated superuser verification for specific commands and modified command syntax recognition.

Added error handling for each one of the newly implemented methods.

Updated sender methods so that they feature multiple database fetching and message sending proceses.